+++ datanode/api/trading_test.go
@@ -167,7 +167,7 @@ func getTestGRPCServer(t *testing.T, ctx context.Context) (tidy func(), conn *gr
 	referralSetsService := service.NewReferralSets(sqlstore.NewReferralSets(sqlConn))
 	teamsService := service.NewTeams(sqlstore.NewTeams(sqlConn))
 	vestingStatsService := service.NewVestingStats(sqlstore.NewVestingStats(sqlConn))
-	referralFeeStatsService := service.NewReferralFeeStats(sqlstore.NewReferralFeeStats(sqlConn))
+	FeesStatsService := service.NewFeesStats(sqlstore.NewFeesStats(sqlConn))
 	fundingPaymentService := service.NewFundingPayment(sqlstore.NewFundingPayments(sqlConn))
 	volumeDiscountStatsService := service.NewVolumeDiscountStats(sqlstore.NewVolumeDiscountStats(sqlConn))
 	volumeDiscountProgramService := service.NewVolumeDiscountPrograms(sqlstore.NewVolumeDiscountPrograms(sqlConn))
@@ -220,7 +220,7 @@ func getTestGRPCServer(t *testing.T, ctx context.Context) (tidy func(), conn *gr
 		referralSetsService,
 		teamsService,
 		vestingStatsService,
-		referralFeeStatsService,
+		FeesStatsService,
 		fundingPaymentService,
 		volumeDiscountStatsService,
 		volumeDiscountProgramService,